chore: move getWalletClient to common folder
