+++ packages/transaction-recovery/src/common/error.ts
@@ -1,0 +1,5 @@
+export const EvmClientError = {
+  WALLET_CLIENT_NOT_FOUND: new Error(
+    "Either 'window.ethereum' or 'privateKey' must be provided"
+  ),
+};
