+++ packages/transaction-recovery/src/evm-add-native-gas/lib/helper.ts
@@ -6,20 +6,12 @@ import {
 } from "@axelarjs/api";
 import { Environment } from "@axelarjs/core";
 
-import "viem/window";
 
 import {
-  createWalletClient,
-  custom,
   Hash,
-  http,
-  publicActions,
   TransactionReceipt,
-  WalletClient,
 } from "viem";
-import { privateKeyToAccount } from "viem/accounts";
 
-import { EvmAddNativeGasError } from "../error";
 import { extractReceiptInfoForNativeGasPaid } from "../lib/getReceiptInfo";
 
 /**
@@ -80,31 +72,3 @@ export async function isInsufficientFeeTx(
   return gmpTx?.gas_status === "gas_paid_not_enough_gas";
 }
 
-/**
- * Get the wallet client to send the native gas payment transaction. If the private key is not provided, then the browser-based wallet (window.ethereum) will be used.
- * @param rpcUrl - The RPC URL of the source chain.
- * @param privateKey - The private key of the sender.
- * @returns The wallet client to send the native gas payment transaction.
- */
-export function getWalletClient(
-  rpcUrl: string,
-  privateKey?: `0x${string}`
-): WalletClient {
-  if (typeof window !== "undefined" && !window.ethereum) {
-    throw EvmAddNativeGasError.WALLET_CLIENT_NOT_FOUND;
-  }
-  if (!privateKey) {
-    throw EvmAddNativeGasError.WALLET_CLIENT_NOT_FOUND;
-  }
-
-  if (privateKey) {
-    return createWalletClient({
-      account: privateKeyToAccount(privateKey),
-      transport: http(rpcUrl),
-    }).extend(publicActions);
-  } else {
-    return createWalletClient({
-      transport: custom(window.ethereum!),
-    }).extend(publicActions);
-  }
-}
